chore: cleans up `Tree` api
